+++ src/ui/buttons.cpp
@@ -1,0 +1,9 @@
+#include <QMainWindow>
+#include "window.h"
+
+void MainWindow::on_loadMore_clicked()
+{
+  loadBookDataChunk(&dataBookBuffer, LOADBEGINNING);
+  bookModel->loadBooks(&dataBookBuffer);
+}
+
