+++ src/ui/window.h
@@ -2,30 +2,10 @@
 #define WINDOW_H
 
 #include <QMainWindow>
-#include <QAbstractTableModel>
-#include "ui_window.h"
+#include "mainTableModel.h"
 extern "C" {
 #include "src/library/library.h"
 }
-
-
-class mainTableModel : public QAbstractTableModel
-{
-  Q_OBJECT
-
-public:
-  explicit mainTableModel(QObject *parent = nullptr);
-  ~mainTableModel();
-
-  int rowCount(const QModelIndex &parent = QModelIndex()) const override;
-  int columnCount(const QModelIndex &parent = QModelIndex()) const override;
-  QVariant data(const QModelIndex &index, int role = Qt::DisplayRole) const override;
-  void loadBooks(Book **data);
-
-private:
-  Book *data_;
-};
-
 
 QT_BEGIN_NAMESPACE
 namespace Ui { class MainWindow; }
